# backend/bait-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API in Laravel",
        "description": "API documentation with Swagger in Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/banners/upload": {
            "post": {
                "tags": [
                    "Banner"
                ],
                "summary": "Upload or replace the authenticated user's banner",
                "description": "Uploads a new banner image for the authenticated user. If a banner already exists, it will be deleted and replaced with the new one.",
                "operationId": "12ee88ff07ba8093d1692ef2512c3b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "banner"
                                ],
                                "properties": {
                                    "banner": {
                                        "description": "The banner image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully uploaded banner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BannerSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    },
                    "422": {
                        "description": "Validation error - invalid or missing file"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/banners/{banner}": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "Get a specific banner",
                "description": "Returns a single banner resource by ID",
                "operationId": "45147c2d2d2d70a997124ee6009411c9",
                "parameters": [
                    {
                        "name": "banner",
                        "in": "path",
                        "description": "ID of the banner to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BannerSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/banners/self": {
            "delete": {
                "tags": [
                    "Banner"
                ],
                "summary": "Delete the authenticated user's banner",
                "description": "Deletes the authenticated user's banner and replaces it with the default banner.",
                "operationId": "2535940fdb80092cf8af33dfe169f573",
                "responses": {
                    "200": {
                        "description": "Banner destroyed and replaced with default",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Banner destroyed. Replaced with default."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/privileged/users/{user}/banner": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Delete a specific user's banner",
                "description": "Deletes the specified user's banner and replaces it with the default banner. Only accessible to admins or moderators.",
                "operationId": "5eb1484e20561bae204c64476ba23c76",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user whose banner will be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner deleted and replaced with default",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Banner for user johndoe destroyed. Replaced with default."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not authorized"
                    },
                    "404": {
                        "description": "User not found or banner not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BannerSchema": {
                "title": "Banner Schema",
                "required": [
                    "id",
                    "url_banners"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url_banners": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/storage/banners/banner1.jpg"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a JWT token to authenticate",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}